requirements.txt
fastapi>=0.70.0
uvicorn>=0.17.0
httpx>=0.22.0
pydantic>=1.9
aiosqlite>=0.17.0
sse-starlette>=1.0.0
python-dotenv>=0.19.0
mcp>=1.0.0
